jumbotron
This is a web-app/feature (prototype) that allows interaction between
the speaker and audience in large seminar or a classroom.

It helps audience/students to raise questions in a stream-lined fashion.
The ability to post questions anonymously helps eliminate any
hindrances.

In the perspective of a speaker, it helps him/her to understand where
the audience have a problem in understanding the delivery. This gives an
opportunity to clarify/answer the questions asked instantaneously and
modify/improvise the delivery in future.

The main intuition behind developing this feature is to enhance the
interactive experience in the classroom environment as a whole.